Refactor CLI display functions into separate module
